backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
exceptiongroup==1.3.0
Flask==3.1.2
flask-cors==6.0.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.182.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.38.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.75.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.2
numpy==2.2.6
pillow==11.3.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
pyparsing==3.2.4
python-dotenv==1.1.1
requests==2.32.5
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.15.3
sniffio==1.3.1
tenacity==9.1.2
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.1.3
